refactor: encapsulate LegacyMetrics
